#5723 A FIX Failover Provider
git-svn-id: http://svn.kolmisoft.com/mor/gui/branches/12@24097 19941b1f-7a74-413c-ab52-0de3e92e9a87

+++ app/models/provider.rb
@@ -68,6 +68,11 @@ class Provider < ActiveRecord::Base
       return false
     end
 
+    if Lcr.count(:conditions => ["failover_provider_id = ?" , self.id]) > 0
+      errors.add(:lcrproviders, _('Cannot_delete_provider_it_is_assigned_as_failover_provider'))
+      return false
+    end
+
     for rule in self.providerrules
       rule.destroy
     end
